productBuilder working. Building 500 products from Kucoin and CoinbasePro. Adding Kraken and Binance later, but we can just pass them in once the SDK's are configured.

+++ src/common/utilities/StrategyManager.js
@@ -134,11 +134,13 @@ class StrategyManager {
   }
 
   deBootShell(obj) {
-    obj.process.kill('SIGHUP')
+    if (obj && obj.process) {
+      obj.process.kill('SIGHUP')
+    }
   }
 
-  pauseAll() {
-    Object.keys(this.state).map(id => this.deBootShell(this.state[id]))
+  async pauseAll() {
+    await Object.keys(this.state).map(id => this.deBootShell(this.state[id]))
   }
 }
